补充新增方法字段

+++ zzxcrm/com.hundsun.crm.biz.core.impl/src/main/java/com/beyond/crm/bean/CrmCustomer.java
@@ -42,8 +42,8 @@ public class CrmCustomer extends AbstractBaseDomain{
 	private String custWebsite;                                  //网址
 	private String custLicenceNo;                                //工商执照号
 	private String custChieftain;                                //法人
-	private Long custBankroll;                                   //资金
-	private Long custTurnover;                                   //年营业额
+	private Double custBankroll;                                   //资金
+	private Double custTurnover;                                   //年营业额
 	private String custBank;                                     //签约银行
 	private String custBankAccount;                              //签约银行账号
 	private String custLocalTaxNo;                               //地税编号
@@ -356,25 +356,25 @@ public class CrmCustomer extends AbstractBaseDomain{
 	/**
 	 * @return the custBankroll
 	 */
-	public Long getCustBankroll() {
+	public Double getCustBankroll() {
 		return custBankroll;
 	}
 	/**
 	 * @param custBankroll the custBankroll to set
 	 */
-	public void setCustBankroll(Long custBankroll) {
+	public void setCustBankroll(Double custBankroll) {
 		this.custBankroll = custBankroll;
 	}
 	/**
 	 * @return the custTurnover
 	 */
-	public Long getCustTurnover() {
+	public Double getCustTurnover() {
 		return custTurnover;
 	}
 	/**
 	 * @param custTurnover the custTurnover to set
 	 */
-	public void setCustTurnover(Long custTurnover) {
+	public void setCustTurnover(Double custTurnover) {
 		this.custTurnover = custTurnover;
 	}
 	/**
